refactor(runs): collapse nine hand-written run-lease fences into one
`AND status = 'running' AND claimed_by = ${owner}` was written out by hand
at nine sites across the worker API, the automation completion path and
the inline gateway path, under four different variable names. #3249 added
a tenth as a shared helper but only the inline callers used it.

One concept spelled nine ways is how a wrong spelling gets in unnoticed:
a fence is invisible when it works and invisible when it silently matches
nothing, so a drifted copy reads exactly like a correct one.

Renames the module to `run-lease` and the helper to `runLeaseFence`,
since worker-held leases fence identically to gateway-inline ones and the
old name claimed otherwise. The embedding-progress write moves to
`runOwnerFence`: it is deliberately status-agnostic, asserting only that
the run is still ours.

No behaviour change — the emitted SQL is identical at every site.
Verified by mutation: dropping the owner term from the shared helper
fails ownership tests at both a pre-existing site
(complete-worker-job-status-guard) and a newly converted one
(heartbeat-lease-guard), so the single definition is load-bearing across
both. 30 tests green across six worker/lease suites.
